move PatchConfig class into its own file
